Extract analysing logic
